refactor: domain -> entity로 변경
