--- studycafe/src/main/java/com/scascanner/studycafe/domain/ReservationStatus.java
@@ -1,5 +1,0 @@
-package com.scascanner.studycafe.domain;
-
-public enum ReservationStatus {
-    RESERVED,CANCELED
-}
