+++ studycafe/src/main/java/com/scascanner/studycafe/domain/entity/reservation/ReservationStatus.java
@@ -1,0 +1,5 @@
+package com.scascanner.studycafe.domain.entity.reservation;
+
+public enum ReservationStatus {
+    RESERVED,CANCELED
+}
